+++ ton_client/src/net/queries.rs
@@ -11,10 +11,15 @@
 * limitations under the License.
 */
 
+use serde_json::Value;
+
 use crate::client::ClientContext;
 use crate::error::{AddNetworkUrl, ClientResult};
-use super::{Error, OrderBy};
-use serde_json::Value;
+use crate::net::{ParamsOfQueryCollection};
+
+use super::Error;
+
+//------------------------------------------------------------------------------------------ query
 
 #[derive(Serialize, Deserialize, ApiType, Clone)]
 pub struct ParamsOfQuery {
@@ -31,25 +36,64 @@ pub struct ResultOfQuery {
     pub result: Value,
 }
 
-#[derive(Serialize, Deserialize, ApiType, Clone)]
-pub struct ParamsOfQueryCollection {
-    /// Collection name (accounts, blocks, transactions, messages, block_signatures)
-    pub collection: String,
-    /// Collection filter
-    pub filter: Option<serde_json::Value>,
-    /// Projection (result) string
-    pub result: String,
-    /// Sorting order
-    pub order: Option<Vec<OrderBy>>,
-    /// Number of documents to return
-    pub limit: Option<u32>,
+/// Performs DAppServer GraphQL query.
+#[api_function]
+pub async fn query(
+    context: std::sync::Arc<ClientContext>,
+    params: ParamsOfQuery,
+) -> ClientResult<ResultOfQuery> {
+    let client = context.get_server_link()?;
+    let result = client
+        .query(&params.query, params.variables, None)
+        .await
+        .map_err(|err| Error::queries_query_failed(err))
+        .add_network_url(client)
+        .await?;
+
+    let result = serde_json::from_value(result)
+        .map_err(|err| Error::queries_query_failed(format!("Can not parse result: {}", err)))
+        .add_network_url(client)
+        .await?;
+
+    Ok(ResultOfQuery { result })
 }
+
+//------------------------------------------------------------------------------- query_collection
 
 #[derive(Serialize, Deserialize, ApiType, Clone)]
 pub struct ResultOfQueryCollection {
     /// Objects that match the provided criteria
     pub result: Vec<serde_json::Value>,
 }
+
+/// Queries collection data
+///
+/// Queries data that satisfies the `filter` conditions,
+/// limits the number of returned records and orders them.
+/// The projection fields are limited to `result` fields
+#[api_function]
+pub async fn query_collection(
+    context: std::sync::Arc<ClientContext>,
+    params: ParamsOfQueryCollection,
+) -> ClientResult<ResultOfQueryCollection> {
+    let client = context.get_server_link()?;
+    let result = client
+        .query_collection(params)
+        .await
+        .map_err(|err| Error::queries_query_failed(err))
+        .add_network_url(client)
+        .await?
+        .clone();
+
+    let result = serde_json::from_value(result)
+        .map_err(|err| Error::queries_query_failed(format!("Can not parse result: {}", err)))
+        .add_network_url(client)
+        .await?;
+
+    Ok(ResultOfQueryCollection { result })
+}
+
+//---------------------------------------------------------------------------- wait_for_collection
 
 #[derive(Serialize, Deserialize, ApiType, Clone, Default)]
 pub struct ParamsOfWaitForCollection {
@@ -69,66 +113,6 @@ pub struct ResultOfWaitForCollection {
     pub result: serde_json::Value,
 }
 
-
-/// Performs DAppServer GraphQL query.
-#[api_function]
-pub async fn query(
-    context: std::sync::Arc<ClientContext>,
-    params: ParamsOfQuery,
-) -> ClientResult<ResultOfQuery> {
-    let client = context.get_server_link()?;
-    let result = client.query(
-        &params.query,
-        params.variables,
-        None,
-    )
-        .await
-        .map_err(|err| Error::queries_query_failed(err))
-        .add_network_url(client)
-        .await?;
-
-    let result = serde_json::from_value(result).map_err(|err| {
-        Error::queries_query_failed(format!("Can not parse result: {}", err))
-    })
-        .add_network_url(client)
-        .await?;
-
-    Ok(ResultOfQuery { result })
-}
-
-/// Queries collection data
-///
-/// Queries data that satisfies the `filter` conditions,
-/// limits the number of returned records and orders them.
-/// The projection fields are limited to `result` fields
-#[api_function]
-pub async fn query_collection(
-    context: std::sync::Arc<ClientContext>,
-    params: ParamsOfQueryCollection,
-) -> ClientResult<ResultOfQueryCollection> {
-    let client = context.get_server_link()?;
-    let result = client.query_collection(
-        &params.collection,
-        &params.filter.unwrap_or(json!({})),
-        &params.result,
-        params.order,
-        params.limit,
-        None,
-    )
-        .await
-        .map_err(|err| Error::queries_query_failed(err))
-        .add_network_url(client)
-        .await?;
-
-    let result = serde_json::from_value(result).map_err(|err| {
-        Error::queries_query_failed(format!("Can not parse result: {}", err))
-    })
-        .add_network_url(client)
-        .await?;
-
-    Ok(ResultOfQueryCollection { result })
-}
-
 /// Returns an object that fulfills the conditions or waits for its appearance
 ///
 /// Triggers only once.
@@ -143,16 +127,45 @@ pub async fn wait_for_collection(
     params: ParamsOfWaitForCollection,
 ) -> ClientResult<ResultOfWaitForCollection> {
     let client = context.get_server_link()?;
-    let result = client.wait_for(
-        &params.collection,
-        &params.filter.unwrap_or(json!({})),
-        &params.result,
-        params.timeout,
-    )
+    let result = client
+        .wait_for_collection(params)
         .await
         .map_err(|err| Error::queries_wait_for_failed(err))
         .add_network_url(client)
         .await?;
 
     Ok(ResultOfWaitForCollection { result })
+}
+
+//--------------------------------------------------------------------------- aggregate_collection
+
+use crate::net::ParamsOfAggregateCollection;
+
+#[derive(Serialize, Deserialize, ApiType, Clone)]
+pub struct ResultOfAggregateCollection {
+    /// Values for requested fields.
+    ///
+    /// Returns an array of strings. Each string refers to the corresponding `fields` item.
+    /// Numeric values is returned as a decimal string representations.
+    pub values: Value,
+}
+
+/// Aggregates collection data.
+///
+/// Aggregates values from the specified `fields` for records
+/// that satisfies the `filter` conditions,
+#[api_function]
+pub async fn aggregate_collection(
+    context: std::sync::Arc<ClientContext>,
+    params: ParamsOfAggregateCollection,
+) -> ClientResult<ResultOfAggregateCollection> {
+    let client = context.get_server_link()?;
+    let values = client
+        .aggregate_collection(params)
+        .await
+        .map_err(|err| Error::queries_query_failed(err))
+        .add_network_url(client)
+        .await?;
+
+    Ok(ResultOfAggregateCollection { values })
 }